server/requirements.txt
websockets>=12.0
pyyaml>=6.0
pathspec>=0.12
pytest>=7.0
pytest-asyncio>=0.21
